Workaround to monocle context menu bug in headless mode

+++ src/test/java/br/com/antoniogabriel/lirelab/acceptance/view/HomeViewTest.java
@@ -18,6 +18,7 @@ import java.util.List;
 import static br.com.antoniogabriel.lirelab.lire.Feature.CEDD;
 import static br.com.antoniogabriel.lirelab.test_utilities.TestPaths.TEST_IMAGES;
 import static br.com.antoniogabriel.lirelab.test_utilities.TestPaths.TEST_ROOT;
+import static br.com.antoniogabriel.lirelab.test_utilities.TestUtils.isHeadless;
 
 public class HomeViewTest extends FXMLTest<HomeFXML> {
 
@@ -122,14 +123,18 @@ public class HomeViewTest extends FXMLTest<HomeFXML> {
 
     @Test
     public void shouldDeleteCollectionByContextMenu() throws Exception {
-        homeView.waitUntilCollectionsAreListed(collection1, collection2);
+        // Nasty hack to execute test only if not in headless mode.
+        // Apparently Monocle cant show ContextMenu in headless mode.
+        if(!isHeadless()) {
+            homeView.waitUntilCollectionsAreListed(collection1, collection2);
 
-        homeView.deleteCollection(collection1);
+            homeView.deleteCollection(collection1);
 
-        homeView.waitUntilCollectionIsNotListed(collection1);
-        homeView.waitUntilCollectionsAreListed(collection2);
+            homeView.waitUntilCollectionIsNotListed(collection1);
+            homeView.waitUntilCollectionsAreListed(collection2);
 
-        COLLECTION_HELPER.createRealCollection(COLLECTION1_NAME, TEST_IMAGES, CEDD);
+            COLLECTION_HELPER.createRealCollection(COLLECTION1_NAME, TEST_IMAGES, CEDD);
+        }
     }
 
     @Test
